requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
SQLAlchemy==2.0.25
PyMySQL==1.1.0
python-jose==3.3.0
passlib==1.7.4
httpx==0.26.0
pandas==2.2.0
reportlab==4.4.10
pymongo==4.16.0
python-dotenv==1.0.0
aiosmtplib==5.1.0
pydantic==2.5.3
bcrypt==4.1.2
cryptography==42.0.0
Pillow==10.2.0
openpyxl==3.1.2
python-multipart==0.0.6
email-validator==2.1.0
